Bass: Broadcast resync helper, out of range case
Helps sinks to resync to external broadcast when out of range.
When there is a SINK PAUSE and BASS cannot sync to the broadcaster,
then it will try to sync till MESSAGE_BROADCAST_MONITOR_TIMEOUT occur,
which is 5 minutes.
If it syncs to the broadcaster then MESSAGE_BROADCAST_MONITOR_TIMEOUT
is removed, but MESSAGE_BIG_MONITOR_TIMEOUT is still counting to the
30 minutes. MESSAGE_BIG_MONITOR_TIMEOUT is used to help sinks to resync
when BIG appear, i.e. when unpausing.

Bug: 363950680
Bug: 363168099
Test: atest BassClientServiceTest
Change-Id: Idee9165dfb756b8eceae92afb0f56a3cc28979fa