Limit gatt clients per app
This change limits the number of GATT clients per app to 32.
This prevents a badly implemented app
from exhausting the global GATT client limit.

The commit also includes a fix for a case where
tBTA_GATTC_CLCB could be accessed after it was deallocated.

This issue was discovered while updating a test in GattClientTest.

Bug: 348559823
Bug: 366473955
Test: atest GattServiceTest GattClientTest

Change-Id: I9c600a867d0c2aee4f92d31fdbf3f8a0775a83c1