BassClientService: Sort scan results for source sync by fails counter
Until now, it was sorted by priority flag and by RSSI. That solution
had the disadvantage that if two broadcasters with the best range,
fails all the time, then any other broadcaster will not be sync.

This change add sorting by sync failure counter between priority
flag and RSSI. Each time when sync fails or lost, counter in increased.
It is cleared when new searching is started.

Bug: 365613358
Bug: 370639684
Test: atest BassClientServiceTest
Change-Id: I8350482009f7a77a5ca02ff981c85b2afb6bcc0b