Encrypt LE link immediately on reconnection
LE link must be encrypted immediately on connection if device are
already bonded.

This is a backport of ag/29056565, but the code needs to go in a
different location because that patch relies on recent feature work.

Ignore-AOSP-First: security
Test: mmm packages/modules/Bluetooth
Bug: 288144143
Merged-In: I607af8214f4fd9c3905f2e053ba136cd6189361f
Change-Id: I8bd0ef5ae1750096fe72feb334022b9895f3b26e