leaudio: Fix switching between LeAudio headsets
When switching between LeAudio Headsets, Java is interested only in
getting ACTIVE groups status with new group.

This is because, LeAudioService is aware about active group and when
switching to new active group, it can update AudioManager in a single
call which improves user experiance when switching during the streaming.

Bug: 369599302
Test: atest LeAudioServiceTest bluetooth_le_audio_client_test
Flag: Exempt, regression tested with unit tests, new test added
Change-Id: Ifc6564e973c1255ad04d5221249b3313fa60d6af