AdapterService: Don't check in teardown
If teardown throw an exception, this is what will be displayed in case
of test failure.
Even if there is an actual & different test failure.

By moving the check within the test, we will now be having the real
failure instead, and triage will be able to be more efficient

Bug: 375091640
Test: atest BluetoothIntrumentationTests
Flag: TEST_ONLY
Change-Id: Icd7117029c99a796ff0b4c18f794935fe9b432c9