Remove lock from FetchAudioProvider.
When binderDiedCallbackAidl is called, it create a new clients and do
the fetching works. This new client is completly independent from the
old client, request still can be sent to the old client.

By removing lock, we prevent BT crash when binderDiedCallbackAidl is
called multiple times, resulting in the lock being destroyed.

Bug: 375051151
Test: mmm packages/modules/Bluetooth
Flag: EXEMPT trivial change
Change-Id: Ie7cc2ba6c2fd9a4c94933677acd5e518b4ec5892