SystemServer: keep gatt apps when scan disallowed
When disallowing "scan while bluetooth is OFF", if bluetooth is ON, it
make no sense to disconnect the apps.
This change will keep the apps connected.

If bluetooth is in ble states, it also doesn't make sense because we are
gonna shut off the bluetooth anyway. So gatt will proceed to the
registration on its own

Bug: 379166793
Bug: 378031425
Test: m . | no test for API user journey yet
Flag: com.android.bluetooth.flags.ble_scan_setting_does_not_disconnect_if_bt_on
Change-Id: I982571e5f0e457cf6ee927393eca01185658c834