Notification: Allow 1 hour window + fix time delta
The alarm that trigger the Bluetooth restart is a low power alarm. It's
wake up time may not be accurate.

To represent that from the notification point of view, we should look to
the date with a window

Also fix the condition for the notification to be displayed after 1 week
and after 6 months

Bug: 348884917
Fix: 348884917
Test: m .
Flag: Exempt trivial
Change-Id: If7e09143f347c6405e5d9259aa7a66101f5cc0ac