leaudio: Improve disconnection during streaming
When LeAudio Set is disconnected during the streaming
from the upper layer, we want to make sure that ACL of both devices are
disconnect when Stream is closed.

Previosly disconnection of first device triggered StopStream but when
second device was disconnected from the upper layer, that cause ACL
disconnection. Usually it worked fine as disconnecting ACL  implies CIS
to be disconnected, but some controllers keeps CIS resources preventing
CIG to be removed later on during the process.

With this CL, all the members will wait until stream is closed

Bug: 381378205
Flag: Exempt, regression verified with unit test
Test: atest bluetooth_le_audio_client_test
Change-Id: I3f531c453ec59da3c91935d73b7163dbc48ae91a